Be more careful to not share memory
The previous code would take the args passed in an logging function call
and combine them with the implied args the logger held (as setup with
With). The problem with the way the code was structured was that we
appended the function call args with to the implied one, which means we
copy them into the shared memory of the logger. That's bad because it
means that the loggers are not goroutine safe in that case.

So rather than deal with any of that, I've refactored the code that
processes the args so that the implied args can be processed separately
from the function call args.