Fix colorization
On unix platforms remove the check for io.File and check if os.Stdout is
a terminal instead of the io.Writer. It is common for programs to wrap
io.Writers, so the output may end up on stdout even if the writer itself
is not a tty.
Removes a panic that would be caused by the above if someone tried to
enable color with a wrapped io.Writer.

On windows remove the panic, and instead keep color off if the writer is
not a tty.