Fix colors not being enabled in some forced contexts.
Turns out that the color package we use was doing it's own "should
colors be on or not" detection, which interferred with our own
detection. This caused situations where colors were forced on to still
be off (and, maddeningly, would show up in weird contexts because it
dependent on Stdout being not a terminal to break when forced).